+++ Sprocket/Graphics/Rendering/Scene3DRenderer.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <Sprocket/Graphics/asset_manager.h>
 #include <Sprocket/Graphics/shader.h>
-#include <Sprocket/Graphics/ShadowMap.h>
+#include <Sprocket/Graphics/shadow_map.h>
 #include <Sprocket/Graphics/buffer.h>
 #include <Sprocket/Scene/ecs.h>
 
@@ -40,6 +40,10 @@ class Scene3DRenderer
 
     Scene3DRenderer(const Scene3DRenderer&) = delete;
     Scene3DRenderer& operator=(const Scene3DRenderer&) = delete;
+
+    void draw_static_meshes(const spkt::registry& registry);
+    void draw_animated_meshes(const spkt::registry& regsitry);
+    void draw_particles(const spkt::registry& registry);
 
 public:
     Scene3DRenderer(asset_manager* assetManager);
